Emissions/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-debug-toolbar==1.7
djangorestframework==3.9.1
feedparser==5.2.1
idna==2.8
pytz==2018.9
requests==2.21.0
schedule==0.6.0
sqlparse==0.2.4
untangle==1.1.1
urllib3==1.24.1
xmltodict==0.11.0
